# distsqlrun: reuse memcontainer memory in sortChunks