@asubiotto 2018-06-21
Review status: :shipit: complete! 0 of 0 LGTMs obtained

---

*[pkg/sql/sqlbase/row_container.go, line 213 at r1](https://reviewable.io/reviews/cockroachdb/cockroach/26874#-LFX_iwB1PZoXaWeP8bU:-LFX_iwB1PZoXaWeP8bV:b-ptnwez) ([raw file](https://github.com/cockroachdb/cockroach/blob/448cb379e55170b5bf51b19621d49086febd015c/pkg/sql/sqlbase/row_container.go#L213)):*
> ```Go
> // the RowContainer to be reused.
> func (c *RowContainer) Clear(ctx context.Context) {
> 	c.chunks = nil
> ```

Do we need to keep this alternate version of `Reset`?

---


*Comments from [Reviewable](https://reviewable.io/reviews/cockroachdb/cockroach/26874)*
<!-- Sent from Reviewable.io -->