@jordanlewis 2018-06-21
Review status: :shipit: complete! 0 of 0 LGTMs obtained

---

*[pkg/sql/sqlbase/row_container.go, line 213 at r1](https://reviewable.io/reviews/cockroachdb/cockroach/26874#-LFX_iwB1PZoXaWeP8bU:-LFXffgxA6mHy-0QN_zM:b-4qicjd) ([raw file](https://github.com/cockroachdb/cockroach/blob/448cb379e55170b5bf51b19621d49086febd015c/pkg/sql/sqlbase/row_container.go#L213)):*
<details><summary><i>Previously, asubiotto (Alfonso Subiotto Marqués) wrote…</i></summary><blockquote>

Do we need to keep this alternate version of `Reset`?
</blockquote></details>

It's a good question. There are several other users of RowContainer.Clear that might actually want this behavior. I haven't taken the time to audit them. But in general, I think both behaviors might be useful depending on the use case...

---


*Comments from [Reviewable](https://reviewable.io/reviews/cockroachdb/cockroach/26874)*
<!-- Sent from Reviewable.io -->