@asubiotto 2018-06-21
Review status: :shipit: complete! 0 of 0 LGTMs obtained

---

*[pkg/sql/sqlbase/row_container.go, line 213 at r1](https://reviewable.io/reviews/cockroachdb/cockroach/26874#-LFX_iwB1PZoXaWeP8bU:-LFXhQTFD1xpBPdRKsI2:b-ux2khc) ([raw file](https://github.com/cockroachdb/cockroach/blob/448cb379e55170b5bf51b19621d49086febd015c/pkg/sql/sqlbase/row_container.go#L213)):*
<details><summary><i>Previously, jordanlewis (Jordan Lewis) wrote…</i></summary><blockquote>

It's a good question. There are several other users of RowContainer.Clear that might actually want this behavior. I haven't taken the time to audit them. But in general, I think both behaviors might be useful depending on the use case...
</blockquote></details>

I don't know when it would actually be beneficial to completely release the memory if you're resetting. To me, the act of resetting implies that you're going to reuse and will have to reallocate what you just released. I would just move the `Reset` behavior into `Clear` but up to you. Am I missing a use case?

---


*Comments from [Reviewable](https://reviewable.io/reviews/cockroachdb/cockroach/26874)*
<!-- Sent from Reviewable.io -->