@RaduBerinde 2018-06-21
Review status: :shipit: complete! 0 of 0 LGTMs obtained

---

*[pkg/sql/sqlbase/row_container.go, line 224 at r2](https://reviewable.io/reviews/cockroachdb/cockroach/26874#-LFXiGhD5wjP5v8EBITM:-LFXlKg_8cpRpW6uuUTN:b-ollvkj) ([raw file](https://github.com/cockroachdb/cockroach/blob/a1091f6db5b545a769a4c604ca61f38e9b614cf7/pkg/sql/sqlbase/row_container.go#L224)):*
<details><summary><i>Previously, RaduBerinde wrote…</i></summary><blockquote>

There is a caveat here - we can overwrite rows previously returned by `At`. This should be a big warning.  Even with the sorter I am not convinced it is safe.. Won't one of these rows end up in a RowChannel? I looked through the code and couldn't find a place where we make a copy (I remember we used to make a copy somewhere).

A safer alternative would be to delete the old rows but keep the rest of the chunks that weren't used yet. Similar to what would happen if we just popped all the old rows.
</blockquote></details>

Ah, we use `copyingRowReceiver` so it is safe.

---


*Comments from [Reviewable](https://reviewable.io/reviews/cockroachdb/cockroach/26874)*
<!-- Sent from Reviewable.io -->